clock: Add basic RClock API.
This is a simple abstraction for retrieving current time and a very simple
method for waiting for a spesific time based on the time given by the clock
itself.

This patch adds both a system clock based on r_time_ts_get_monotonic and
a test clock for testing purposes.